test/e2e/framework move func logClusterImageSources to test/e2e/e2e.go
